==> A_train/0_test.ipynb <==
import os
import sys
sys.path.append('..')
import numpy as np
import pandas as pd
from library import config, load_o_data

df_test = load_o_data.load_df_test()

df_trade = load_o_data.load_df_trade()

df_test_trade = pd.merge(df_test[['core_cust_id', 'prod_code', 'deal_date']], df_trade, on=['core_cust_id', 'prod_code'], how='left', suffixes=('', '_y'))
df_test_trade = df_test_trade.loc[df_test_trade['deal_date'] > df_test_trade['deal_date_y'], ]
df_test_trade_g = df_test_trade.groupby(['core_cust_id', 'prod_code', 'deal_date']).agg({'trade_id': len}).reset_index()
df_test_trade_g

df_test_predict = pd.merge(df_test, df_test_trade_g, on=['core_cust_id', 'prod_code', 'deal_date'], how='left')
df_test_predict['y'] = [1-int(x) for x in pd.isnull(df_test_predict['trade_id'])]
df_test_predict

df_test_predict[['id', 'y']].to_csv(os.path.join(config.result_p_dir, 'his_20210109.csv'), index=False)

np.mean(df_test_predict['y'])
